+++ projects/journey-maps-client/src/lib/components/marker-details/components/teaser/teaser.component.ts
@@ -22,6 +22,7 @@ import {LocaleService} from '../../../../services/locale.service';
 })
 export class TeaserComponent implements OnInit {
 
+  @Input() shouldRender: boolean;
   @Input() selectedMarker: Marker;
   @Input() template?: TemplateRef<any>;
   @Output() closeClicked = new EventEmitter<void>();